chore: rename util files
better reflects their purpose
